# stream-hub-api/app/Infrastructure/Models/Video.php
<?php

namespace App\Infrastructure\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Factories\HasFactory;

class Video extends Model
{
    use HasFactory;

    protected $fillable = ['title', 'created_at', 'category_id', 'hls_path', 'description', 'thumbnail', 'site_id', 'views', 'likes'];

    public function site()
    {
        return $this->belongsTo(Site::class);
    }

    public function category()
    {
        return $this->belongsTo(Category::class);
    }
}