# stream-hub-api/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Descrição da minha API",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get list of videos",
                "operationId": "af0d3163c12f615e3e97f9dc9472b3df",
                "responses": {
                    "200": {
                        "description": "A list of videos"
                    }
                }
            }
        },
        "/api/videos/{id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get a single video by ID",
                "operationId": "4f3140b090445a904b4598c1004542bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video details"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Videos"
                ],
                "summary": "Atualiza os dados do vídeo",
                "operationId": "23ffd9a9a03141709ccdbd88fc782080",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "likes": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Novo título"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vídeo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Video": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "hls_path": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string"
                    },
                    "site_id": {
                        "type": "integer"
                    },
                    "views": {
                        "type": "integer"
                    },
                    "likes": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Videos",
            "description": "Videos"
        }
    ]
}